+++ src/s21_auxfunct.h
@@ -14,7 +14,7 @@ union DateDecLL {
 
 int checkDecimal(s21_decimal numb);
 void CheckFloat8(float src, int * mantissa_seven, int * scale_seven);
-double mDecimal(s21_decimal numb); 
+long double mDecimal(s21_decimal numb); 
 void printBits(const void *ptr, size_t size);
 void printCharBits(char num);
 void printFloatBits(float num);
